hw-6 getById delete separate
