+++ spring-filg/spring-hw-06/src/main/java/ru/otus/spring/repositories/BookRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface BookRepository {
-    Optional<Book> getBookById(long id);
+    Optional<Book> getBookById(long bookId);
 
     List<Book> getBooksByAuthor(Author author);
