+++ spring-filg/spring-hw-06/src/main/java/ru/otus/spring/repositories/jpa/AuthorRepositoryJpa.java
@@ -19,7 +19,6 @@ public class AuthorRepositoryJpa implements AuthorRepository {
 
     @Override
     public List<Author> getAuthors() {
-
         return em.createQuery("select a from Author a", Author.class).getResultList();
     }
