+++ spring-filg/spring-hw-06/src/main/resources/application.yml
@@ -14,11 +14,20 @@ spring:
       settings:
         web-allow-others: true
 
+#  jpa:
+#    generate-ddl: false
+#    hibernate:
+#      ddl-auto: none
+##      ddl-auto: validate
+#
+#    show-sql: true
   jpa:
     generate-ddl: false
     hibernate:
       ddl-auto: none
-#      ddl-auto: validate
+    properties:
+      hibernate:
+        format_sql: true
 
     show-sql: true
